+++ src/server/api/user.js
@@ -1,6 +1,7 @@
 import { mapFieldsToModel } from './lib/utils'
 import { r, User } from '../models'
 import { rolesAtLeast } from '../../lib/permissions'
+import { userOrgsWithRole } from '../models/cacheable_queries'
 
 export function buildUserOrganizationQuery(queryParam, organizationId, role) {
   const roleFilter = role ? { role } : {}
@@ -34,15 +35,7 @@ export const resolvers = {
       if (!user || !user.id) {
         return []
       }
-      let orgs = r.knex.select('organization.*')
-        .from('organization')
-        .join('user_organization', 'organization.id', 'user_organization.organization_id')
-        .where('user_organization.user_id', user.id)
-      if (role) {
-        const matchingRoles = rolesAtLeast(role)
-        orgs = orgs.whereIn('user_organization.role', matchingRoles)
-      }
-      return orgs.distinct()
+      return userOrgsWithRole(role, user.id)
     },
     roles: async(user, { organizationId }) => (
       r.table('user_organization')
